+++ neuronquant/utils/logger.py
@@ -37,7 +37,6 @@ log_format = u'[{record.time:%Y-%m-%d %H:%M}] {record.channel}::{record.level_na
 
 def get_nestedlog(level='DEBUG', file='quantrade.log', uri=None):
     # Default uri: tcp://127.0.0.1:5540
-    print ' ******************** Level asked for: ', level
     if uri is not None:
         log_setup = NestedSetup([
             ZeroMQHandler(uri),